Split up process_constraints.py
